requirements.txt
pdf2image
pytesseract
openai
flask
flask-cors
flask-login
flask-sqlalchemy
werkzeug
gunicorn
psycopg2-binary
easyocr
pypdfium2
numpy
pillow